refactoring to make deprecations clear and start with a clean slate with Chris' variation on things
