+++ geminidr/interactive/controls.py
@@ -18,8 +18,14 @@ from abc import ABC, abstractmethod
 
 from bokeh.events import PointEvent
 
+__all__ = ["controller", "Controller"]
 
-""" This is the active controller.  It is activated when it's attached figure sees the mouse enter it's view.  """
+
+""" This is the active controller.  It is activated when it's attached figure sees the mouse enter it's view.
+
+Controller instances will set this to listen to key presses.  The bokeh server will use this to send keys
+it recieves from the clients.  Everyone else should leave it alone!
+"""
 controller = None
 
 
@@ -41,22 +47,21 @@ class Controller(object):
     control to the `Controller`.  The `Tasks` are also able to update the help
     text to give contextual help.
     """
-    def __init__(self, gifig, aperture_model, band_model, helptext):
+    def __init__(self, fig, aperture_model, band_model, helptext):
         """
         Create a controller to manage the given aperture and band models on the given GIFigure
 
         Parameters
         ----------
-        gifig : :class:`GIFigure`
+        fig : :class:`~Figure`
             plot to attach controls to
-        aperture_model : :ckass:`GIApertureModel`
+        aperture_model : :class:`GIApertureModel`
             model for apertures for this plot/dataset, or None
         band_model : :class:`GIBandModel`
             model for bands for this plot/dataset, or None
         helptext : :class:`Div`
             div to update text in to provide help to the user
         """
-        self.gifig = gifig
         self.helptext = helptext
         self.tasks = dict()
         if aperture_model:
@@ -68,13 +73,13 @@ class Controller(object):
         self.y = None
         # we need to always know where the mouse is in case someone
         # starts an Aperture or Band
-        gifig.figure.on_event('mousemove', self.on_mouse_move)
-        gifig.figure.on_event('mouseenter', self.on_mouse_enter)
-        gifig.figure.on_event('mouseleave', self.on_mouse_leave)
+        fig.on_event('mousemove', self.on_mouse_move)
+        fig.on_event('mouseenter', self.on_mouse_enter)
+        fig.on_event('mouseleave', self.on_mouse_leave)
 
-        self.sethelptext()
+        self.set_help_text()
 
-    def sethelptext(self, text=None):
+    def set_help_text(self, text=None):
         """
         Set the text in the help area.
 
@@ -174,11 +179,11 @@ class Controller(object):
         if self.task:
             if self.task.handle_key(key):
                 self.task = None
-                self.sethelptext()
+                self.set_help_text()
         else:
             if key in self.tasks:
                 self.task = self.tasks[key]
-                self.sethelptext(self.task.helptext())
+                self.set_help_text(self.task.helptext())
                 self.task.start(self.x, self.y)
 
     def handle_mouse(self, x, y):
@@ -204,7 +209,7 @@ class Controller(object):
         if self.task:
             if self.task.handle_mouse(x, y):
                 self.task = None
-                self.sethelptext()
+                self.set_help_text()
 
 
 class Task(ABC):
