+++ geminidr/interactive/deprecated/__init__.py
@@ -1,0 +1,4 @@
+# Maybe a bit extreme, but I just want to keep all this around for a bit longer while
+# also making it easy for people to ignore it.  The path forward will be more like
+# what Chris did in refactoring traceApertures.  New generic fitting code will go in
+# the fit package.
